Handle prompt from Muse, transition to Drawing
This *works*, in terms of the state transitions and such,
but there's a lot I want to refactor.
The state machine shouldn't be implicitly contained in Room,
which is getting too many fields and methods.

Also, the transition functions shouldn't do anything but update states.
It'd be better to just let them return an error instead of running them
in a separate thread.