Implement scrolling for Chat component
* Don't auto scroll on load
* Do keep to bottom if user has scrolled to bottom
* Anchor form to bottom

In other words, if the user is scrolled to bottom,
they'll see new messages as they appear.
If not, they stay on the focused message.
User must scroll to bottom once.

Main issue I see is that Firefox autohides scrollbar
when not scrolling, so it's not obvious to user that
scrolling is necessary.
Unfortunately, I don't know if there's a way around this
in CSS instead of Firefox's about:config.

https://superuser.com/questions/1720362/firefox-scroll-bar-disappearing